Added blocks and grid controls
