--- Assets/Materials/Grid.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 11f6f1799a99a6749b7cd379d96d0488
-timeCreated: 1435168908
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
