--- Assets/Materials/GridOverlay.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e01b4da85ff9d534f96d1fd94594d674
-timeCreated: 1444541093
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
