+++ Assets/Resources/Blocks/CarvedStone.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 178878982f7487e4291e042f9564456f
+folderAsset: yes
+timeCreated: 1458512185
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
