+++ Assets/Resources/Blocks/CarvedStone/CarvedStoneBlock.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class CarvedStoneBlock : Block {
+	public CarvedStoneBlock() {
+		name = "CarvedStone";
+		displayName = "Carved Stone";
+	}
+}
