+++ Assets/Resources/Blocks/CarvedStone/CarvedStoneBlock.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0ec44d2fe342dc84792407fb1b106b90
+timeCreated: 1458512294
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
