+++ Assets/Resources/Blocks/CarvedStone/CarvedStoneMaterial.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e7461a24fc6dea45909a6416b45f86a
+timeCreated: 1458512305
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
