+++ Assets/Resources/Blocks/Pillar.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 085c66b0128efe147bd123aad5617a98
+folderAsset: yes
+timeCreated: 1458513886
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
