+++ Assets/Resources/Blocks/Pillar/PillarBlock.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class PillarBlock : Block {
+	public PillarBlock() {
+		name = "Pillar";
+		displayName = "Pillar";
+	}
+}
