+++ Assets/Resources/Blocks/Pillar/PillarBlock.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c7977dcfd5c5dfc46a2fd4b9f6e9eceb
+timeCreated: 1458513895
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
