+++ Assets/Resources/Blocks/Pillar/PillarMaterial.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e39f8d72e932de4ba4d03927a95f36b
+timeCreated: 1458513904
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
