--- Assets/Scripts/Camera.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 086c50f53c6370f47bbb660959c291fc
-folderAsset: yes
-timeCreated: 1434841482
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
