--- Assets/Scripts/Camera/CameraOrbit.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 1ecabcc535772f14fab39496376e3453
-timeCreated: 1434841463
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
