--- Assets/Scripts/Camera/GridRenderer.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a6c7d13f2e1a3ee4da46a5c139fdf187
-timeCreated: 1435167059
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
