+++ Assets/Scripts/Design/CameraMove.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: da3dd7f0cbdea17489c527cf245c52a2
+timeCreated: 1434843082
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
